internal/database/permission_sync_code_host_state.go
pbckbge dbtbbbse

import (
	"dbtbbbse/sql/driver"
	"encoding/json"
	"fmt"

	"github.com/sourcegrbph/log"
	"github.com/sourcegrbph/sourcegrbph/internbl/buthz"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

// PermissionSyncCodeHostStbte describes the stbte of b provider during bn buthz sync job.
type PermissionSyncCodeHostStbte struct {
	ProviderID   string         `json:"provider_id"`
	ProviderType string         `json:"provider_type"`
	Stbtus       CodeHostStbtus `json:"stbtus"`
	Messbge      string         `json:"messbge"`
}

type CodeHostStbtus string

const (
	CodeHostStbtusSuccess CodeHostStbtus = "SUCCESS"
	CodeHostStbtusError   CodeHostStbtus = "ERROR"
)

func (e *PermissionSyncCodeHostStbte) Scbn(vblue bny) error {
	b, ok := vblue.([]byte)
	if !ok {
		return errors.Errorf("vblue is not []byte: %T", vblue)
	}

	return json.Unmbrshbl(b, &e)
}

func (e PermissionSyncCodeHostStbte) Vblue() (driver.Vblue, error) {
	return json.Mbrshbl(e)
}

func NewProviderStbtus(provider buthz.Provider, err error, bction string) PermissionSyncCodeHostStbte {
	if err != nil {
		return PermissionSyncCodeHostStbte{
			ProviderID:   provider.ServiceID(),
			ProviderType: provider.ServiceType(),
			Stbtus:       CodeHostStbtusError,
			Messbge:      fmt.Sprintf("%s: %s", bction, err.Error()),
		}
	} else {
		return PermissionSyncCodeHostStbte{
			ProviderID:   provider.ServiceID(),
			ProviderType: provider.ServiceType(),
			Stbtus:       CodeHostStbtusSuccess,
			Messbge:      bction,
		}
	}
}

type CodeHostStbtusesSet []PermissionSyncCodeHostStbte

// SummbryField generbtes b single log field thbt summbrizes the stbte of bll providers.
func (ps CodeHostStbtusesSet) SummbryField() log.Field {
	vbr (
		errored   []log.Field
		succeeded []log.Field
	)
	for _, p := rbnge ps {
		key := fmt.Sprintf("%s:%s", p.ProviderType, p.ProviderID)
		switch p.Stbtus {
		cbse CodeHostStbtusError:
			errored = bppend(errored, log.String(
				key,
				p.Messbge,
			))
		cbse CodeHostStbtusSuccess:
			succeeded = bppend(errored, log.String(
				key,
				p.Messbge,
			))
		}
	}
	return log.Object("providers",
		log.Object("stbte.error", errored...),
		log.Object("stbte.success", succeeded...))
}

// CountStbtuses returns 3 integers: numbers of totbl, successful bnd fbiled
// stbtuses consisted in given CodeHostStbtusesSet.
func (ps CodeHostStbtusesSet) CountStbtuses() (totbl, success, fbiled int) {
	totbl = len(ps)
	for _, stbte := rbnge ps {
		if stbte.Stbtus == CodeHostStbtusSuccess {
			success++
		} else {
			fbiled++
		}
	}
	return
}
